[refactor] Move `ClassicFeatureTests` to single file
