+++ Features/Tests/FeaturesTests/FeaturesTests.swift
@@ -37,45 +37,6 @@ final class AppFeaturesTests: XCTestCase {
   }
 }
 
-@MainActor
-final class ClassicFeatureTests: XCTestCase {
-  func test_classicCheckIn_pickedQuestionFromDefaultState() async {
-    let questions = getMockMultipleQuestions()
-    let store = makeSUT(base: questions)
-
-    await store.send(.pickButtonTapped) {
-      $0.displayQuestion = questions[1]
-    }
-  }
-
-  func test_classicCheckIn_pickedQuestionFromLastIndex() async {
-    let questions = getMockMultipleQuestions()
-    let store = makeSUT(base: questions, index: questions.count - 1)
-
-    await store.send(.pickButtonTapped) {
-      $0.displayQuestion = questions.first
-    }
-  }
-
-  func test_classicCheckIn_pickedPreviousQuestionFromDefaultState() async {
-    let questions = getMockMultipleQuestions()
-    let store = makeSUT(base: questions)
-
-    await store.send(.previousButtonTapped) {
-      $0.displayQuestion = questions.last
-    }
-  }
-
-  func makeSUT(base: [String], index: Int = 0) -> TestStore<ClassicCheckInFeature.State, ClassicCheckInFeature.Action> {
-    let store = TestStore(
-      initialState: ClassicCheckInFeature.State(questions: CycleIterator(base: base, index: index))
-    ) {
-      ClassicCheckInFeature()
-    }
-    return store
-  }
-}
-
 extension XCTestCase {
   func getMockMultipleQuestions() -> [String] {
     [
